Add files via upload
An S Ranked Mission
Secret agents Pytachi and Csame from Akatsuki Federation, are out on an S Ranked Mission for the second time. Their mission is to steal The Scroll of Seals from Konohagakure, also known as the village hidden in the leaves. The scroll is kept in a room having a hi-tech security system in the Hokage’s Mansion. The room has only one door which opens only on entering the correct password. If someone tries to break in the room forcefully, or enters a wrong password, all the dead Hokages will be summoned using the reanimation Jutsu instantly, who will protect The Scroll of Seals. All these Hokages were the most skilled and powerful Ninjas of their times, and hardly anyone can fight all of them at once.

Csame is a highly skilled martial artist and has an extraordinary capability of performing various tasks in a flash. He is the only one in the whole world who can match the skills of all the Hokages. But still, he might not be able to defeat all of them and steal the scroll all alone. Hence, he teamed up with Pytachi for this mission. Pytachi is a natural genius, having extraordinary perception and deduction skills. But, Pytachi has a disease called Alzheimer’s (A progressive disease that destroys memory and other important mental functions).

The agents have managed to reach the door of the room without being noticed. But now, they need to crack the password. In a state of high nervousness, Pytachi forgot that he has Alzheimer's and remembered the procedure of deriving the password which he deduced when they tried to steal the scroll for the first time.

He told the the procedure to Csame:

2 integers X and Y are written on the door, these are the hints to derive the password, which represent the range [X,Y].
Let Cmin be the minimum count of factors of any number in this range.
SCmin is the sum of all those numbers in the range [X,Y] which have Cmin number of factors.
Let Cmax be the maximum count of factors of any number in this range.
SCmax is the sum of all those numbers in the range [X,Y] which have Cmax number of factors.
Let Csum be the sum of counts of factors of those numbers in range [X,Y] for which the count of factors is neither Cmax nor Cmin.
S′ is the sum of all the numbers whose count of factors was added in Csum, i.e. the sum of all the numbers in the range [X,Y] for which the count of factors is neither Cmax nor Cmin.
The English Alphabets are indexed, starting from ′a′ at index 0.
F=alphabet at index (Cminmod26).
M=alphabet at index (Csummod26)
L=alphabet at index (Cmaxmod26).
An alphabet must be in uppercase if its index is Even, and in lowercase if its index is Odd.
The password contains exactly 5 fields in this format (without any spaces between them): Alphabet Integer Alphabet Integer Alphabet
These 5 fields of the password are F, SCmin, M, SCmax and L respectively.
At this moment, Csame appreciated Pytachi by saying “You're Amazing Pytachi! I cannot believe the you have Alzheimer’s.”

This made Pytachi remember that he has Alzheimer’s and he was suddenly blank for a moment. And, as usual, Csame derived the password within a flash using the above procedure, and entered it. But, as soon as he submitted the password, the security system was triggered and the Hokages were summoned, and an epic battle started between the agents and the Hokages.

Csame and Pytachi somehow managed to escape from Konohagakure, but both of them were badly injured. While they were under treatment, Pytachi told Csame that he forgot to tell one rule for the password, the rule was:

If S′ is odd, then the actual password is the reverse of the derived password.
Also, he mentioned that the values of X and Y were different this time from the ones that he saw on their first mission.

This mission was a failure, but the Akatsuki Federation now knows the procedure to derive the password. They want to use this knowledge in a perfect way, they believe that there might be more traps inside the room too, so they cannot rely on anyone but Csame and Pytachi. Hence, they will wait till both of them recover completely.

Until then, Akatsuki Federation is planning to make sure that this procedure derives the correct password. They will do T tests. In each test, they will send bots which will be programmed to calculate and enter the password. These bots will just enter the password then close the door if it opens. They cannot take the risk of trying to steal the scroll using these bots, because if the security system is triggered due to any other trap inside the room, then of course the mission will fail, but the current Hokage will also realize that the password is cracked. Hence, he might change the password derivation strategy if this happens. This will waste all the efforts of Pytachi and Csame.

They want the bots to find the password as fast as Csame, so they have challenged your team to implement the above procedure most efficiently, to prove your prowess in programming.

Input
First line contains an integer T, the number of tests.
Each test is described in a single line containing two space-separated integers denoting the values of X and Y.
Output
For each test, print a single line containing the password.
Constraints
1≤T≤5
1≤X≤Y≤1012
1≤Y−X+1≤5∗105
Sample Input
2
8 13
5 100
Sample Output
G21l42C
M204E5501C
Explanation
For testcase 1:

Factors of 8: {1,2,4,8}, i.e total 4 factors.
Factors of 9: {1,3,9}, i.e. total 3 factors.
Factors of 10: {1,2,5,10}, i.e. total 4 factors.
Factors of 11: {1,11}, i.e. total 2 factors.
Factors of 12: {1,2,3,4,6,12}, i.e. total 6 factors.
Factors of 13: {1,13}, i.e. total 2 factors.
According to the representations in problem statement, here:

Cmin=2
SCmin=11+13=24
Cmax=6
SCmax=12
Csum=4+3+4=11
S′=8+9+10=27
Cminmod26=2mod26=2. Hence, F=alphabet at index 2= ′c′
Csummod26=11mod26=11. Hence, M=alphabet at index 11= ′l′
Cmaxmod26=6mod26=6. Hence, L=alphabet at index 6= ′g′
Since the index of F (i.e. 2) is even, F must be in uppercase, i.e. F= ′C′
Since the index of M (i.e. 11) is odd, M must be in lowercase, i.e. M= ′l′
Since the index of L (i.e. 6) is even, L must in uppercase, i.e. L= ′G′
The password is of the form F SCmin M SCmax L , hence the password is C24l12G.
Since S′ (i.e. 27) is odd, the correct password is the reverse of the derived password. Hence, the correct password is G21l42C.